requirements.txt
torch
torchvision
numpy
Pillow
PyQt5
tqdm